+++ lib/core.dart
@@ -3,5 +3,6 @@ library core;
 export 'enums.dart';
 export 'extensions.dart';
 export 'helpers.dart';
+export 'instance.dart';
 export 'models.dart';
 export 'utils.dart';